Switch from recursion to iteration for uniq()
As the number of unique files accessed grow in number, the recursive
implementation of uniq() can lead to a segfault because of stack size
limitations. One can work around this by increasing the stack size limit
(ulimit -s), however that is not always possible.

This commit changes the implementation of uniq() to use iteration so
that such limits do not lead to problems.