Use memfd_create() to create shared memory region on Linux.
These Linux-specific memory regions are significantly better than
/dev/shm ones since they are automatically reclaimed when all
file descriptors pointing to them are closed. Moreover, they do
not have a real filesystem name, so they cannot conflicts when
multiple commands with the same key are launched.

For context, see this bug entry for the Fuchsia build which is
experiencing fsatrace-related issues:

  https://bugs.fuchsia.dev/p/fuchsia/issues/detail?id=95360